python-api/requirements.txt
fastapi
uvicorn[standard]
nanonets-ocr
python-multipart